Several SAC-related changes
Changed the computation of the log probability density, changed where the actions boundary is for SAC reasons (not including the action conversions), added a minimum samples so that the replay buffer is not full of untrained policy experiences, slightly changed configuration to ensure better stability of V-values and Q-values by reducing the discount factor. Also modified the models a little to ensure consistency.